+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Comparing T_EOC between reactive and non-reactive caused a `ValueError`
+- Add `copy` argument to `AltExperiment` class initializer
 
 ### Changed